+++ pythainlp/benchmarks/__init__.py
@@ -5,6 +5,11 @@
 
 __all__: list[str] = [
     "BleuScore",
+    "CharLevelStats",
+    "GlobalStats",
+    "RougeScore",
+    "TokenizationStats",
+    "WordLevelStats",
     "benchmark",
     "bleu_score",
     "character_error_rate",
@@ -14,9 +19,16 @@ __all__: list[str] = [
 
 from pythainlp.benchmarks.metrics import (
     BleuScore,
+    RougeScore,
     bleu_score,
     character_error_rate,
     rouge_score,
     word_error_rate,
 )
-from pythainlp.benchmarks.word_tokenization import benchmark
+from pythainlp.benchmarks.word_tokenization import (
+    CharLevelStats,
+    GlobalStats,
+    TokenizationStats,
+    WordLevelStats,
+    benchmark,
+)